requirements.txt
asgiref==3.2.3
autopep8==1.4.4
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.13.2
cloudinary==1.19.1
dj-database-url==0.5.0
Django==3.0
django-cors-headers==3.2.0
django-rest-multiple-models==2.1.3
djangorestframework==3.10.3
djangorestframework-simplejwt==4.4.0
mock==3.0.5
psycopg2==2.8.4
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
python-decouple==3.3
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
